Move mapoverlay styles to separate file
