+++ src/components/shared/MapInformationOverlay/MapInformationOverlay.js
@@ -1,86 +1,11 @@
 // @flow
 
 import React, { Component, type Node } from "react";
-import {
-  Dimensions,
-  ScrollView,
-  StyleSheet,
-  Text,
-  Image,
-  TouchableOpacity,
-  View,
-} from "react-native";
-import { Colors, TextStyles } from "../../styles/";
-import { StyleSheetUtils } from "../../utils/";
-import SharingService from "../../services/SharingService";
+import { ScrollView, Text, Image, TouchableOpacity, View } from "react-native";
+import styles from "./style";
+import SharingService from "../../../services/SharingService";
 
-const closeIcon = require("../../images/ic_close.png");
-
-const screenWidth = Dimensions.get("window").width;
-const screenHeight = Dimensions.get("window").height;
-
-const defaultMargin = 17;
-const closeButtonSize = 26;
-const scrollViewMaxHeight = screenHeight - 138; // magic number here, but roughly (listitem + header + margins)
-
-const styles = StyleSheet.create({
-  container: {
-    position: "absolute",
-    width: screenWidth - defaultMargin * 2,
-    margin: defaultMargin,
-    backgroundColor: Colors.white,
-    shadowColor: Colors.shadow,
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowRadius: 5,
-    shadowOpacity: 1,
-  },
-  scrollView: {
-    maxHeight: scrollViewMaxHeight,
-  },
-  scrollViewNoTitle: {
-    maxHeight: scrollViewMaxHeight,
-    width: "90%",
-  },
-  closeButtonContainer: {
-    position: "absolute",
-    top: 10,
-    right: 10,
-  },
-  closeButton: {
-    width: closeButtonSize,
-    height: closeButtonSize,
-    backgroundColor: Colors.white,
-    borderRadius: closeButtonSize / 2,
-  },
-  titleText: StyleSheetUtils.flatten([
-    TextStyles.title,
-    {
-      color: Colors.black,
-      fontSize: 24,
-      lineHeight: 36,
-      marginTop: 22,
-      marginHorizontal: defaultMargin,
-    },
-  ]),
-  descriptionText: StyleSheetUtils.flatten([
-    TextStyles.description,
-    {
-      margin: defaultMargin,
-      color: Colors.greyBodyText,
-      lineHeight: 22,
-    },
-  ]),
-  shareContainer: {
-    paddingLeft: defaultMargin,
-    paddingRight: defaultMargin,
-    paddingBottom: 4,
-    flexDirection: "row",
-    justifyContent: "flex-end",
-  },
-});
+const closeIcon = require("../../../images/ic_close.png");
 
 type TrailInformation = {
   title: ?string,
