+++ src/components/shared/TrailView/index.js
@@ -4,7 +4,7 @@ import React from "react";
 import { TouchableWithoutFeedback, View } from "react-native";
 import styles from "./style";
 import MapWithListView, { type MapItem } from "../../shared/MapWithListView";
-import MapInformationOverlay from "../../shared/MapInformationOverlay";
+import MapInformationOverlay from "../../shared/MapInformationOverlay/MapInformationOverlay";
 import AudioPlayerView from "../AudioPlayerView";
 import DownloadButtonContainer from "../DownloadButton";
